Extract inference row runner and stream wire adapter from scheduler.
Move tier-job orchestration into inference_row_runner and domain-to-NDJSON
mapping into transport/inference_stream_wire so the scheduler stays a generic
queue executor and payload builders are public API.
