+++ packages/api/api/analytics/military_score_inference/inference_stream_session.py
@@ -1,0 +1,44 @@
+"""Per-row inference NDJSON stream session state."""
+
+from __future__ import annotations
+
+import queue
+import uuid
+from dataclasses import dataclass, field
+
+from api.analytics.military_score_inference.inference_cancel import InferenceCancelToken
+from api.analytics.military_score_inference.inference_stream_domain_events import (
+    InferenceStreamDomainEvent,
+)
+from api.analytics.military_score_inference.inference_stream_orchestration import (
+    InferenceStreamOrchestration,
+)
+from api.analytics.military_score_inference.inference_stream_scope import InferenceStreamScope
+from api.analytics.military_score_inference.models import InferenceObservation
+from api.analytics.military_score_inference.policy_ladder import PolicyLadderState
+from api.models.game import TurnInfo
+
+
+@dataclass
+class InferenceRowStreamSession:
+    """Per-row NDJSON stream state shared between the request thread and workers."""
+
+    player_id: int
+    observation: InferenceObservation
+    turn: TurnInfo
+    game_id: int
+    perspective: int
+    turn_number: int
+    cancel_token: InferenceCancelToken = field(default_factory=InferenceCancelToken)
+    event_queue: queue.Queue[InferenceStreamDomainEvent] = field(default_factory=queue.Queue)
+    run_id: str = field(default_factory=lambda: str(uuid.uuid4()))
+    ladder_state: PolicyLadderState | None = None
+    orchestration: InferenceStreamOrchestration | None = None
+
+    @property
+    def stream_scope(self) -> InferenceStreamScope:
+        return InferenceStreamScope(
+            game_id=self.game_id,
+            perspective=self.perspective,
+            turn_number=self.turn_number,
+        )
